Fix installation

+++ includes/functions.sh
@@ -167,7 +167,9 @@ EOF
 			fi
 			echo "$(textb [INFO]) - Setting your hostname..."
 			if [[ -f /lib/systemd/systemd ]]; then
-				apt-get update -y > /dev/null 2>&1 && apt-get install dbus > /dev/null 2>&1
+				if [[ -z $(dpkg --get-selections | grep -E "^dbus.*install$) ]]; then
+					apt-get update -y > /dev/null 2>&1 && apt-get install dbus > /dev/null 2>&1
+				fi
 				hostnamectl set-hostname $sys_hostname
 			else
 				echo $sys_hostname > /etc/hostname
